% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blblm.R
\name{blblm}
\alias{blblm}
\title{Preform bootstrapping linear regression}
\usage{
blblm(formula, data, parallelization = FALSE, m = 10, B = 5000)
}
\arguments{
\item{formula}{A formula}

\item{data}{A data frame}

\item{parallelization}{A boolean value on whether or not you want to use parallelization}

\item{m}{An int that is the upper bound of indexes of your data}

\item{B}{An int that tells the function how many times you want to bootstrap}
}
\value{
res An list of the bootstrapped models
}
\description{
Preform bootstrapping linear regression
}
