% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{coef.blblm}
\alias{coef.blblm}
\title{Get the coefficients of the model}
\usage{
\method{coef}{blblm}(object, parallelization = FALSE, ...)
}
\arguments{
\item{object}{A model}

\item{parallelization}{A boolean value on whether or not you want to run parallelization}
}
\value{
the coefficients of the model
}
\description{
Get the coefficients of the model
}
