% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint.R
\name{confint.blblm}
\alias{confint.blblm}
\title{Get the confidence interval for the model}
\usage{
\method{confint}{blblm}(object, parallelization = FALSE, parm = NULL, level = 0.95, ...)
}
\arguments{
\item{object}{A model}

\item{parallelization}{A boolean value for whether or not you want to run parallelization}

\item{level}{A double that is the confidence interval level}
}
\value{
out A list of confidence intervals for each variable
}
\description{
Get the confidence interval for the model
}
