% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.blblm}
\alias{predict.blblm}
\title{Get predictions for the model}
\usage{
\method{predict}{blblm}(
  object,
  parallelization = FALSE,
  new_data,
  confidence = FALSE,
  level = 0.95,
  ...
)
}
\arguments{
\item{object}{A model}

\item{parallelization}{A boolean for whether or not you want to run parallelization}

\item{new_data}{A dataframe of data you want to predict}

\item{confidence}{A boolean of whether or not you want a confidence interval}

\item{level}{A double that is the confidence interval}
}
\value{
prediction values
}
\description{
Get predictions for the model
}
