% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigma.R
\name{sigma.blblm}
\alias{sigma.blblm}
\title{Get the sigma value for the model}
\usage{
\method{sigma}{blblm}(object, parallelization = FALSE, confidence = FALSE, level = 0.95, ...)
}
\arguments{
\item{object}{A model}

\item{parallelization}{A boolean value on whether or not you want to use parallelization}

\item{confidence}{A boolean value on whether or not you want a confidence value}

\item{level}{A double that is the confidence level you want}
}
\value{
sigma A double
}
\description{
Get the sigma value for the model
}
